move interfaces and edit imports for remaining interfaces
