--- src/components/api/const/interfaces/profileUpdateInterface.tsx
@@ -1,11 +1,0 @@
-export interface UpdateProfileInterface {
-  url?: string;
-  bio?: string;
-  venueManager?: boolean;
-}
-
-export type UpdatedProfileData = {
-  avatar?: { url: string };
-  bio?: string;
-  venueManager?: boolean;
-};
