+++ src/components/api/const/interfaces/singleVenueInterfaces.tsx
@@ -56,3 +56,26 @@ export interface BookingCalendarProps {
   onDateChange: (fromDate: string | null, toDate: string | null) => void;
   venue: SingleVenueResponse;
 }
+
+export interface bookingTileInterface {
+  venue: VenueTileInterface;
+}
+
+interface VenueTileInterface {
+  id: string;
+  location: {
+    city?: string;
+    country?: string;
+  };
+  maxGuests: number;
+  media: Media[];
+  meta: {
+    wifi: boolean;
+    parking: boolean;
+    breakfast: boolean;
+    pets: boolean;
+  };
+  name: string;
+  price: number;
+  rating: number;
+}
